### 2017-06-03 #121 @adamhsparks
bomrang
_bomrang_ provides functions to interface with Australian Government Bureau of Meteorology (BOM) data, fetching data and returning a tidy data frame of précis forecasts, current weather data from stations or ag information bulletins.

```
Package: bomrang
Title: Fetch Australian Government Bureau of Meteorology Data
Version: 0.0.4-1
Authors@R: c(person("Adam", "Sparks", role = c("aut", "cre"),
    email = "adamhsparks@gmail.com"),
    person("Hugh", "Parsonage", role = "aut",
    email = "hugh.parsonage@gmail.com"),
    person("Keith", "Pembleton", role = "aut",
    email = "keith.pembleton@usq.edu.au"))
Description: Provides functions to interface with Australian Government Bureau
    of Meteorology (BOM) data, fetching data and returning a tidy data frame of
    précis forecasts, current weather data from stations or agriculture
    bulletin data.
    curl,
    foreign,
    httr,
    tools,
    xml2
    covr,
NeedsCompilation: no
```

https://github.com/ToowoombaTrio/bomrang

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.): data retrieval

-   Who is the target audience?  Scientists that work with meteorological data from BOM or anyone interested in working with BOM data in R or other applications more easily than the formats BOM provides to the public

There are a few other R packages that I've found that provide access BOM data or ways to import it into R.

- _BOMdataRipper_, https://github.com/johnDorian/BOMdataRipper, provides the most similar functionality for BOM data as _bomrang_ but is not yet on CRAN and development seems sporadic. It only fetches station data, it does not provide the ag bulletins or précis forecasts like _bomrang_. It uses the older _XML_ and _RCurl_ packages, whereas _bomrang_ uses the newer _xml2_ and _curl_ packages to take advantage of more efficient R packages. Additionally, _BOMdataRipper_ provides functionality for visualising the data. _bomrang_ takes a much more UNIX like approach only doing one thing, downloading data and serving it to the end-user in an R `data.frame` object.

- _bomdata_, https://github.com/mbertolacci/bomdata, this package provides only  rainfall data and site metadata only in a SQLite database.

- _ReadAxfBOM_, https://github.com/MarkusLoew/ReadAxfBOM, this package reads Axf files from BOM but does not automate the downloading, only importing from local
files and does not offer forecasts or ag bulletins.

- _bomr_, https://github.com/njtierney/bomr, this package is the basis for much of _bomrang_. It was set up for the #AUUnconf2016, which I attended. We didn't end up with a package but a vignette that is referenced in the _bomrang_ README file. Several of the ideas presented here could be incorporated into _bomrang_.

The rOpenSci packages _GSODR_ and _rnoaa_ provide similar functionality for differing data sets. There is some overlap with data in _GSODR_ and _rnoaa_ (GHCN data), but the audience and end use will probably differ in most cases.

Lastly, _bomrang_ is just a cool name. Not that I can take credit for it or that it affects the package functionality, but I'll just point it out anyway.

    - [x] The package is deposited in a long-term repository with the DOI: http://doi.org/10.5281/zenodo.598301

- [x] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings: None

- [x] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? Please describe any exceptions: Yes, no exceptions

  - Brooke Anderson, @geanders
  - Michael Sumner, @mdsumner
