### 2017-06-06 ToowoombaTrio/bomrang#19 @HughParsonage
jsonlite preferred over rjson
I chose `library(rjson)` arbitrarily but it seems `jsonlite` is explicitly preferred by rOpenSci (https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#tools) and they're probably right.
