### 2017-06-06 ToowoombaTrio/bomrang#19 @adamhsparks
OK. Thanks, I'll update the review request with this issue so they know we're addressing it.
