### 2017-06-06 #121 @adamhsparks
We're aware that _jsonlite_ is preferred over _rjson_ by rOpenSci and are addressing that.

https://github.com/ToowoombaTrio/bomrang/issues/19
