### 2017-06-07 ToowoombaTrio/bomrang#19 @adamhsparks
thanks.
