### 2017-06-07 #121 @adamhsparks
We've addressed this, _bomerang_ now uses `jsonlite`

https://github.com/ToowoombaTrio/bomrang/commit/fdee7bd517d77b4d1dfc701c32432d5926693bbe
