### 2017-06-21 #121 @karthik
### Editor checks:

#### Editor comments

Thanks for your submission, @adamhsparks!  I'm currently looking  for reviewers. 🙏

In the meantime here is the output of `goodpractice::gp()` that can help you and the reviewers. Note that reaching 100% code coverage isn't compulsory, the output just encourages you to write more of them.

── GP bomrang ──────────────────────────────────────────────────────────────────

It is good practice to

    R/get_ag_bulletin.R:44:NA
    R/get_ag_bulletin.R:46:NA
    R/get_ag_bulletin.R:49:NA
    R/get_ag_bulletin.R:52:NA
    R/get_ag_bulletin.R:53:NA
    ... and 553 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.

    R/get_ag_bulletin.R:126:1
    R/get_current_weather.R:121:1
    R/get_current_weather.R:129:1
    R/get_precis_forecast.R:124:1
    R/internal_functions.R:7:1
    ... and 10 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘tools’ All declared Imports should be used.
────────────────────────────────────────────────────────────────────────────────
