### 2017-06-22 #121 @adamhsparks
Hi @karthik, thanks for your suggestions. A few comments and some questions for discussion.

- I've removed _tools_ and it passes checks, but when I had checked this locally and on WinBuilder I didn't get any NOTES, i.e. _tools_ not being used? How did you find that?

- Coverage is better than it appears using `gp()` since it's not possible to run the tests on CRAN to download files, Travis takes ~20 minutes to run, but coverage is 99% according to [covr.io](https://codecov.io/gh/ToowoombaTrio/bomrang)

- What is the thought on long lines? I've reduced line lengths everywhere in code as a general practices except where it's a comment for the user. Should these be broken just to fit the code window with indentation? Change indentation? Something else?

- I've also imported only the functions necessary for bomrang to work, not as a whole
