### 2017-06-22 ropensci/bomrang#21 @mpadge
Add 9am & 3pm state bulletins
Adam, Hugh, and Keith, I'm excited about reviewing this package, and will take this opportunity to confess that i worked as a BoM forecaster for a couple of years so consider myself well-suited for the job. The first thing I wondered on checking the package out was why it didn't extract the daily (09:00, 15:00) weather bulletins? These data represent the key archived observations for the BoM on which more things depend than any other set of data. They'd be easy to extract, so ... could you add these please?

[Example here](http://www.bom.gov.au/vic/observations/9am_bulletin.shtml)

Things to consider:
1. This appears to me a fairly striking omisson that you could readily address prior to me submitting my review, but if you prefer to keep the review process free of complications, this issue could simply remain until post-review. (There may also be more to come.)
2. I've often wondered, and don't really know, what the status would be of somebody archiving themselves these daily obs? The BoM ultimately charges for dissemination of the full historical data set, so there may be some kind of infringement issue that might need to be considered?
