### 2017-06-22 ToowoombaTrio/bomrang#21 @adamhsparks
@mpadge,
Thanks for your enthusiasm. Originally this started as a script for another project @Keith-Pembleton  and I are working on, https://github.com/ToowoombaTrio/WINS, which only needed the forecast as we were drawing the observed data from SILO. I added the ag bulletin since it was also in an XML format.

@HughParsonage kindly added the current weather after finding the package.

I'll confess, I'd not really considered adding the 9am & 3pm state bulletins, but we can certainly look into it. Are they only a webpage as you've linked to? No json or xml interface?

About the infringement issues, I'm not keen to archive any data with the package, only the metadata about the stations themselves, and provide facilities to retrieve and parse the weather data. But we can check with BoM to be safe. Right now I figured we were since we're retrieving what they are putting out as json and xml...
