### 2017-06-22 ToowoombaTrio/bomrang#21 @HughParsonage
Extracting the 9am and 3pm state bulletins directly would be difficult and I think should be out-of-scope for this package. The data is already captured in `get_current_weather`, and I foresee that maintaining the function to extract the unstructured page in perpetuity would be a task. Although obviously important and central, that page is intended for internet browsers -- i.e. human, not machine, readers. We could just read the data and print it to the console line-by-line, but that seems not much better than just printing the URL.

On the archiving question, your point is interesting and the BoM seems silent on its copyright page. My guess would be it would be very similar to Google's API Terms of Service: you can access the API we provide, you can even charge others to access it, but you can't store the data (or, obviously, compete with the Bureau by selling the data).
