### 2017-06-22 ToowoombaTrio/bomrang#21 @mpadge
@HughParsonage The last thing I want to do is start the review process off on a bad note, but ... the bulletins represent the single major twice-daily data output of the entire operation that is the BoM, and a package called `bomrang` should surely incorporate this? Current weather is not the same, because different stations have differen reporting schedules, yet the only two things they must all do are the 9 & 15 obs. These are also the most detailed obs of the day, and the bulletins are the only way that the BoM disseminates a national weather snapshot.

As for extracting it: it's a simple html table, very easy to scrape, and it looks like it would be very cleanly structured. The format hasn't really ever changed since ... the invention of the internet. Loads of packages do data scraping of far messier sites. My own  [`bikedata` package](https://github.com/mpadge/bikedata) has to pull in some pretty messy data, and includes checks to ensure appropriate fails when and if `html` or other data structures change. It really would be quite simple, and ought to be able to be done with really very little code.

An alternative: I'd be perfectly happy to help post-review, if you'd prefer just to leave this issue until then. If you were to simply give confirmation that you/we would do it down the track, then i'd be happy to simplly state that it's a non-issue coz you're working on it.
