### 2017-06-22 ToowoombaTrio/bomrang#21 @adamhsparks
@mpadge, I think your proposition that you help after the review would be the most ideal. I understand @HughParsonage's objection to scraping a webpage, I have the same reservations as well, but based on your experience I'm willing to give it a go. My first thought was that _xml2_ could likely parse the page since it is XHTML and it's what I used for the précis and ag bulletin.

I also know that BoM is working on updating the APIs to access the data, who knows when that will come to fruition though. But I've requested to be an early tester before they are released to the general public. So I guess there is no guarantee that the current observations or the forecasts and bulletins will remain in the same format either. This isn't any different.
