### 2017-06-23 ToowoombaTrio/bomrang#21 @adamhsparks
I had a chat with @Keith-Pembleton today about this. As I understand it, BoM data should be available through SILO already under Creative Commons, but isn't yet. The plan is in place, but it's just not yet available. That's going to be a well structured data set that we can use for this in the future when that's actually in place.
