### 2017-06-24 #121 @mpadge
No contribbution guidelines; otherwise all okay

##### Paper (for packages co-submitting to JOSS)

The package contains a `paper.md` with:

I note only that four levels of versioning might be overkill - current version is `0.0.4-1`. Up to the authors, but with just 3 levels you'd now be up to `0.0.11` (or maybe `0.2.x`), which still allows an approximately infinite future before version numbers get too high, and makes it easier for others to track version numbers.

---

This package certainly provides a unique and demonstrably important functionality of allowing ready access to data from the Australian Bureau of Meteorology (BoM). It's very cleanly coded and the authors are to be commended for ensuring the `tidy`ness of all `data.frame` objects returned from the primary functions.

Note that as a former meteorologist for the BoM turned producer of `R` data packages, this idea has long been on my to-do/wish list. Accordingly, please do not interpret the following comments as overt criticisms, rather as reflections of my desire to see this package become the best it can! I'm merely a bit envious that you guys got around to it before i did, and came up with such a great package name for this task.

-----

#### Loading package

No problems installing from a local tarball, but problems arose after installing using `devtools::install_all()`

devtools::install_github("ToowoombaTrio/bomrang",lib="~/R/x86_64-pc-linux-gnu-library/3.4")
```
This installs okay, but then:
update_forecast_locations()
# Overwriting existing database
# Error in bzfile(file, "wb") : cannot open the connection
# In addition: Warning message:
# In bzfile(file, "wb") :
#   cannot open bzip2-ed file '<local_dirs>/bomrang/inst/data/AAC_codes.rda',
#   probable reason 'No such file or directory'
```
This arises largely because the file paths for internal data are not specified correctly. The following instances exemplify:
1. The line (#34) `destfile = paste0(tempdir(), "AAC_codes.dbf")` is wrong because it won't insert the necessary system delimiter. **All** args should simply be passed to `file.path`, and not `paste0`. At it stands, this creates the `.dbf` file one directory up from `tempdir()`, with a name that is a mash of the last part of `tempdir()` plus the desired file name, so the function works but not really as intended. Critically, it creates a file **outside** the `tempdir()` of the current env, so will fail on any systems on which users do not have permission to do that!
2. Same for the subsequent lines (#39-40)
   AAC_codes <-
        foreign::read.dbf(paste0(tempdir(), "AAC_codes.dbf"), as.is = TRUE)
   ```
   This successfully reads back the file, but **not** from `tempdir()`, rather from one dir up.
3. Same again for L#47-48
    path <- file.path(file.path(pkg, "data"), paste0("AAC_codes.rda"))
   ```
   this should just be
    path <- file.path(file.path(pkg, "data"), "AAC_codes.rda")
   ```
4. But Step#3 won't work in a failsafe way because, for example, using `devtools::load_all()` generates
   system.file(package = "bomrang")
   # [1] "<local>/bomrang/inst"
   ```
   and
   list.files(system.file(package = "bomrang"))
   # [1] "CITATION" "paper"
   ```
   this does not contain the expected `/data` directory. Installing tarball or using `devtools::install_github()` returns the expected result (that is, the full `pkg` with `data/` directory).

5. Same issues will also affect `update_station_locations()`

All of this ought to be able to be resolved by simply  using robust commands to access the internal package data, such as:
system.file ("data", "AAC_codes.rda", package = "bomrang")
```
This just works as is for all of the installation methods described above, but also leads to another suggestion

#### package data

In line with the general principle of exporting as little as possible for package functionality, i suspect it would be better to put all of this data in `inst/data` so it's not lazy-loaded with the package. Will the users really need access to "AAC_codes" or "JSONurl_latlon_by_station_name"? Putting these in `inst/data` will make the package considerably tidier.

Also note that "JSONurl_latlon_by_station_name" only just passes `lint` tests for variable names which should be <= the precisely 30 chars this is. Could it be shorter?

-----

#### package functionality

All functions perform as described and intended, and the `data.frame` tidying is exemplary. Given the amount of work the Toowoomba Trio (also a great name, by the way!) put into tidying, I wonder whether it might not be worth while returning all `data.frame` objects as `tibble` objects? This would offer a couple of distinct advantages:
1. Many functions, such as `get_precis_forecast()`, return very large dfs.  Naively examining these quickly fills the screen with junk, and requires a doubling of much code in the vignette such as,
   QLD_forecast <- get_precis_forecast(state = "QLD")
   head(QLD_forecast)
   ```
   If it were a tibble, it would suffice to simply have,
   get_precis_forecast(state = "QLD")
   ```
2. At the moment,
   class (stations_site_list)
   # [1] "tbl_df"     "tbl"        "data.frame"
   ```
   yet all other objects are `data.frame` objects. I suspect this particular data is worth keeping in `data/` (rather than `inst/data`), yet having one `tibble` and `data.frame` objects for everything else is perhaps inconsistent (even untidy, one might assert).

Happy to leave the decision of whether to do this to the trio.

-----

### Package functions

#### `get_precis_forecast`

Given the nice way you've used `agrep` for partial matching of `get_current_weather()`, why not use that for state names? I should be able to enter just one letter for all states except NSW and NT which would require two.
vic <- get_precis_forecast(state = "VIC") # okay
vic <- get_precis_forecast(state = "vic") # okay
vic <- get_precis_forecast(state = "vi") # Not okay
vic <- get_precis_forecast(state = "victoria") # Not okay - really!!?
```

#### `get_current_weather`

w <- get_current_weather("castle")
# UNIVERSITYCASTLEMAINE PRISONCASTERTON'?
```
Note that the recommended `station_name` values are not separated by spaces, which they should be - please fix. Then note that no warning is generated for
w <- get_current_weather("castlemaine")
w <- get_current_weather("castlema")
```
yet one is nevertheless generated for the *unambiguous*
w <- get_current_weather("castlem")
```
Then note that,
w <- get_current_weather("castlema")
w$name [1]
# [1] "Castlemaine"
```
and yet the warning told me `station_name = 'CASTLEMAINE PRISON'`. I think there needs to be more consistency in handling the names of observation stations and what they are referred to in the `data.frame` objects, and personally suspect that the **full** names ought to be used at all times - so this ought to be "CASTLEMAINE PRISON" at all times. This is likely to be particularly important because many stations within a given town/city/area have changed location over time, and second denominators often serve to clarify which station is being used. ("CASTLEMAINE PRISON" is, for example, relatively new, with previous observations at "CASTLEMAINE" actually taken at a different location.)

Now note the nomenclature for the names of weather stations,
names (stations_site_list) # it's called "name"
names (get_current_weather("castle"))
# it's called "name", but the `Warning` refers to `station_name`
names (get_precis_forecast(state = "vic")) # it's called "location"
```
The variable names which hold these stations also ought to be consistently named.

Digging a bit further
w <- get_current_weather("castlema")
cbind (w$lat, w$lon)
#       [,1]  [,2]
# [1,] -37.1 144.2
# [2,] -37.1 144.2
# [3,] -37.1 144.2
```
and yet,
w <- get_current_weather(latlon = c (-37, 144))
#    144.2392
```
Why are the locations rounded in the `data.frame`? Surely it would be better to keep the full values?

On a positive note: The conversion to local user times is fantastic and likely to be highly useful and informative - great idea!

### `get_ag_bulletin`

ag_bulletin <- get_ag_bulletin(state = "tassie")
```
Oh yeah, that doesn't work. Okay, try this,
ag_bulletin <- get_ag_bulletin(state = "tasmania")
#   TASMANIAis not recognised as a valid state or territory
```
What?! But actually, I just reiterate that point to show again a missing white space ("TASMANIAis") that needs to be fixed. Okay,
ag_bulletin <- get_ag_bulletin(state = "tas")
head (ag_bulletin)
```
Now these locations are called `"station"`, so we've now had locations variously referred to in four different ways:
1. `"name"`
2. `"station_name"`
3. `"location"`
4. `"station"`

Please fix!

Finally, the following code re-iterates some of the points raised above:
agb <- get_ag_bulletin(state = "vic")
agb [which (agb$station == "Kyabram"), ]
# 5  NA  NA  NA NA
get_current_weather (latlon = c (-36.3, 145)) [1, ]
#   1           NA           NA      -       -     CALM            0           0
```
1. The warning `"Using station_name = "` is misleading because it's here called `"name"`
2. The `lat` and `lon` values from `get_ag_bulletin()` are full precision, yet again truncated in `get_current_weather()`

#### `lint` and `goodpractise` notes

I note here that the `goodpractise` notes regarding long lines mostly refer to lines such as this (from `get_ag_bulletin()`:
```
stop(
  "\nThe server with the bulletin files is not responding. Please retry again later.\n"
))
```
This would be better written,
stop(paste0("\nThe server with the bulletin files is not responding.",
            "Please retry again later.\n"))
```
This is neater and easier to read, and will not raise any `lint` notes. There are many other instances of lines of code being written like this (from `update_forecast_locations()`):
AAC_codes <-
foreign::read.dbf(paste0(tempdir(), "AAC_codes.dbf"), as.is = TRUE)
```
which would be more neatly formatted as,
AAC_codes <- foreign::read.dbf(paste0(tempdir(), "AAC_codes.dbf"),
                               as.is = TRUE)
```
(Although of course that line is wrong for reasons mentioned above.) And note that you can also use `#nolint` at the end of problematic lines to avoid `lint`/`goodpractise` notes that really are unavoidable.

-----

#### A more general concern

`bomrang` claims to return "current weather data", yet omits the 9am and 3pm state bulletins. These are the major daily data output of the BoM (more recent incorporation of hydrological monitoring notwithstanding) and serve as the authoritative historical repository for Australian climate data. I really do believe that the `bomrang` package ought to offer these data as well, as discussed in [this `bomrang` issue](https://github.com/ToowoombaTrio/bomrang/issues/21), in response to which the authors have indicated their willingness to incorporate these data within the package.

Finally, I would expect some kind of message on package load describing the source of the data and acknowledging BoM. (Not blowing my own trumpet here, it's just one i know, but see the equivalent for ['OpenStreetMap' data](https://github.com/ropensci/osmdata/blob/master/R/zzz.R#L12-L16).)
