### 2017-06-24 #121 @adamhsparks
Thanks for a very thorough review @mpadge

I'll open your suggestions as issues and start working on addressing them.

A few thoughts I had reading your review.

- I'd had a message on loading regarding BoM, but later dropped it after reading discussions on some of the mail lists expressing a general distaste for them.

- Yesterday I started moving my packages over to using `file.path` in place of `paste0` and you've just given me my first demonstrated reason why it's not just faster to operate but necessary in some cases, thank you.

- I debated on returning a `tibble`. A discussion on Twitter had support for both returning a `tibble` and a simple `data.frame` as _bomrang_ now does. In the end, based on the discussion, I decided to return a `data.frame` because new users might not be familiar with how to convert back to a `data.frame` but someone wanting a `tibble` will certainly know how to create one from the _bomrang_ output. I'm not strongly convinced either way I suppose and can be convinced that a `tibble` is the appropriate format.
