### 2017-06-24 #121 @adamhsparks
Regarding station and location names.

With the précis forecast, my understanding was that these are for "locations" not for a given station but rather a forecast for a named location usually a post office or something of that nature. If this is incorrect, I'll rename the field to stations.
