### 2017-06-25 #121 @mpadge
Oh yeah, I had kinda forgotten that. Precis forecasts are actually first made for "districts" (not exactly sure of that terminology), then disaggregated to select "towns". [Example here](http://www.bom.gov.au/vic/forecasts/towns.shtml). The verbal forecast gets written for the district, then numbers for each select town in a defined district. For example, if you click on all "Northern Country" towns, you'll see that some get their own slightly different verbal forecast, while others just inherit the generic "Northern Country Area" forecast.

What does this mean for your nomenclature? I guess you ought to use "towns" for the precis forecasts, after first ensuring that all states do refer to them as towns. It would also be useful to explain these nuances in the function doc.
