### 2017-06-27 ToowoombaTrio/bomrang#31 @adamhsparks
Return stations_site_list as a data frame
Renamed after discussion below.

Will simply ensure that `stations_site_list` is a data.frame and that all functions include option to return `data.table` object as `get_current_weather()` now does.

---------------------------------------------------------

From @mpadge in https://github.com/ropensci/onboarding/issues/121

> All functions perform as described and intended, and the data.frame tidying is exemplary. Given the amount of work the Toowoomba Trio (also a great name, by the way!) put into tidying, I wonder whether it might not be worth while returning all data.frame objects as tibble objects? This would offer a couple of distinct advantages:

> Many functions, such as get_precis_forecast(), return very large dfs. Naively examining these quickly fills the screen with junk, and requires a doubling of much code in the vignette such as,
`QLD_forecast <- get_precis_forecast(state = "QLD")`
`head(QLD_forecast)`
>If it were a tibble, it would suffice to simply have,
`get_precis_forecast(state = "QLD")`
>At the moment,
`class (stations_site_list)`
`# [1] "tbl_df"     "tbl"        "data.frame"`
> yet all other objects are data.frame objects. I suspect this particular data is worth keeping in data/ (rather than inst/data), yet having one tibble and data.frame objects for everything else is perhaps inconsistent (even untidy, one might assert).
