### 2017-06-27 ToowoombaTrio/bomrang#31 @adamhsparks
@HughParsonage, I'm thinking this is OK to do, especially with the whole tidyverse and the fact that we do return a tidy data frame.

Looking at `get_current_weather()` gives me the idea to offer the default return as a `tibble` but optionally return a `data.table` for the other objects that `bomrang` returns.

Thoughts?
