### 2017-06-27 ToowoombaTrio/bomrang#31 @HughParsonage
I'm quite OK with either `tibbles` or `data.frames` by default (I'd personally use `data.table`.)

If we both had infinite time and will, I think the 'correct' way to do would be provide an option `options("bomrang.print.method" = "tibble")` and let the user's preference dictate the print method.

But it seems to be a mostly cosmetic criticism of the package. Feel free to assign me work if you'd prefer me to do it.
