### 2017-06-27 ToowoombaTrio/bomrang#31 @adamhsparks
Yeah, it does seem mostly cosmetic. I'm still on the fence. However, if nothing else we should make sure the package is standardised, `stations_site_list` is a `tibble` at the moment.

I'd initially settled on using `data.frame` just for the end user to make the call thinking that someone who wanted a tibble would know how to use `foo <- as_tibble(get_ag_bulletin(state = "Qld"))` and could make that call themselves.

I'm happy to return a `data.frame` and offer `data.table` as an option for all the returns for consistency across the package.
