### 2017-06-27 ToowoombaTrio/bomrang#31 @mpadge
My two cents worth: I see it really just as a cosmetic decision, and the review comment was really just about inconsistency given the package claim to return a "tidy data frame" of everything. It seems fair to expect "tidy" to mean consistent. Easiest path would just be converting the current sole `tibble` to a simple `data.frame` and job would be done. `tibble`-ifying everything does have the further disadvantage that the `print` method truncates full view to (usually) 5 columns only, whereas `print` for `data.frame`s at least lets you see all columns.
