### 2017-06-27 ToowoombaTrio/bomrang#31 @adamhsparks
Thanks @mpadge, I think we're on the same page then fro your post and my
last.

I'll detibble the lone tibble and we should offer the same options in all
functions, i.e. to return a data.frame as well

--
Adam H Sparks
