### 2017-06-27 ToowoombaTrio/bomrang#31 @adamhsparks
For consistency, should we make `JSONurl_latlon_by_station_name` a `data.frame` object as well and convert it to a `data.table` object in the functions when fetching current weather?
