### 2017-06-27 ToowoombaTrio/bomrang#31 @mpadge
But wouldn't this be better as internal data? Then it doesn't really matter what it is. No user is ever going to need to access that, right?
