### 2017-06-29 ToowoombaTrio/bomrang#31 @adamhsparks
@mpadge, @HughParsonage  and I discussed exposing the data externally here:
https://github.com/ToowoombaTrio/bomrang/issues/23

We decided to leave data exposed. I can see possible use cases similar to what Hugh already described there, for myself as well. It also means that you can update only one data set at a time. If it's internal it's all one Rda file, unless I'm mistaken?
