### 2017-06-29 ToowoombaTrio/bomrang#31 @HughParsonage
> But it seems to be a mostly cosmetic criticism of the package. Feel free to assign me work if you'd prefer me to do it.
