### 2017-07-12 ToowoombaTrio/bomrang#22 @adamhsparks
Fixed in https://github.com/ToowoombaTrio/bomrang/commit/73abda361c5fa70434df7f19a88b281fd4557de4
