### 2017-07-12 ToowoombaTrio/bomrang#31 @adamhsparks
I've come around and agree with @mpadge. There's no need to pollute the user-space with the data. I've started moving the data to be internal only. Two of the three databases currently are internal only. I'll update the third shortly.

If someone really wants the data, they can refer to the Rmd files in /data-raw https://github.com/ToowoombaTrio/bomrang/tree/devel/data-raw to create their own.
