### 2017-07-13 ToowoombaTrio/bomrang#31 @adamhsparks
Closed with https://github.com/ToowoombaTrio/bomrang/commit/d9f13cbe73f840f787b941e70e67e10f782a9a68
