### 2017-07-14 #121 @geanders
## Package Review

- [ x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

- [x ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [ x] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [x ] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [ x] **Function Documentation:** for all exported functions in R help
- [ x] **Examples** for all exported functions in R Help that run successfully locally
- [ x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`). (This package has this in `CONDUCT.md`, referenced in the `README`)

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [ x] **A short summary** describing the high-level functionality of the software
>- [ x] **Authors:**  A list of authors with their affiliations
>- [x ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [ x] **Installation:** Installation succeeds as documented.
- [ x] **Functionality:** Any functional claims of the software been confirmed.
- [ x] **Performance:** Any performance claims of the software been confirmed.
- [x ] **Automated tests:** Unit tests cover essential functions of the package
- [ x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

Estimated hours spent reviewing: 3

---
