### 2017-07-14 #121 @geanders
### Review Comments

I think this is a great package, which will potentially be very useful to some researchers, with a very, very cool name. There are some ideas in here that are really cool, like the ability to update the data that shipped with the package through the two `update_*` functions. The documentation is very helpful, particularly the warning messages, which typically provide enough information to help a new or infrequent R user to troubleshot as they use the package. I did not find any critical problems in the code beyond those pointed out by the first reviewer, but I did have some minor suggestions throughout.

### Code

- I'm not sure if this has the potential to cause any problems for some users, but it might be worthwhile to try to revise the code to use some of the recommended techniques for standard evaluation when programming with tidyverse packages rather than getting around CRAN notes by defining column names in dataframes as `NULL` within functions (i.e., all the spots noted now with `# CRAN NOTE avoidance`). I know that the underscore alternative dplyr functions can do that (`mutate_` rather than `mutate`), but I think they've been deprecated now. I think the latest best practices are outlined here: http://dplyr.tidyverse.org/articles/programming.html
- As noted by the first reviewer, `JSONurl_latlon_by_station_name` is a very long variable name. It may be worth renaming this dataset to something shorter.
- Some of the package dependencies have gone through a number of big changes since they were first released. It may be worthwhile adding minimum version requirements in the DESCRIPTION file, in case some users haven't updated packages in a while.
- Some of the code could probably be refactored to be much shorter. One example is the series of checks for whether certain columns are in the data frame within the `.get_obs` function. Another example is the code to create `xmlbulletin` based on the state chosen in a couple of the functions, which could probably be shortened a good bit using the `case_when` function from `dplyr`. Although, in that second specific case,  given that the number of states should stay the same in the future, I guess there probably shouldn't be any maintanance problems resulting from the current approach for that piece of code.
- A couple of the functions (`cook` and `.get_obs`) are defined within other functions. Is this for some scoping-related reason? Otherwise, it might make the code a bit more readable to define them outside of other functions.
- In some places, the code switches between tidyverse and base R for data cleaning. It may be easier to read / maintain the code if it sticks to tidyverse instead. (Again, not a critical issue, just something to consider to make maintenance easier.)
- I really like `%notin%` in the internal functions and will be using that idea in my daily life (usually I do `!(... %in% ...)`, which gets annoying).
- For `get_precis_forecast`, `return(tidy_df)` seems to be repeated. I couldn't tell why (maybe just a typo?).

### Documentation

- Throughout, I found the documentation very helpful. I was particularly impressed with the efforts taken to make the warning and error messages clear and useful to users.
- Some of the warning messages need a space added when they're pasted with variables (I think the other reviewer also noted this).
- I agree with the other reviewer's suggestion to combine the vignettes into one. Also, I would recommend evaluating more of the code in the vignette, so that users can see what the output looks like.
- The processing of raw data was well documented through use of .Rmd files in the `data-raw` folder.
- It might be nice to include a map of current station locations in the vignette, so users can see the typical coverage of weather stations.
- I thought the datasets in the package were also well-documented.
- For the `latlon` argument, it would be helpful to add what conventions the user should use for latitude and longitude (for example, are units decimal degrees? Should southern latitudes be negative, or is that assumed?)
- It's very nice that some functions try to find a match to a station name if the user didn't get the station name exactly right (and how it gives a message about other possible matches in that case).
- There's only one "big idea" in the package that I would suggest increasing documentation on. The `update_*` functions are a cool idea, but I think it would be useful to explain a bit more what's going on with these functions in the vignette. In particular, it would be good if users understand exactly where (in their computer's file system) these updates are being saved. Also, it may be helpful for reproducibility to add some comments on the vignette about how this functionality might result in users with the same version of the package may end up with different results from running the same code (i.e., if one has updated the data at some point and one has not or updated at a different time).
