### 2017-07-15 #121 @adamhsparks
Thanks @geanders!

Good thoughts and comments. I like your suggestions here and we'll work on getting them incorporated.
