### 2017-07-15 ToowoombaTrio/bomrang#34 @adamhsparks
Add map of station locations to vignette
From https://github.com/ropensci/onboarding/issues/121

> "It might be nice to include a map of current station locations in the vignette, so users can see the typical coverage of weather stations."
