### 2017-07-15 ToowoombaTrio/bomrang#34 @adamhsparks
Heh we both worked on this today.

https://github.com/ToowoombaTrio/bomrang/commit/5ec5052f48079ebf589a3ccdd47d080a349ea303
