### 2017-07-15 ToowoombaTrio/bomrang#34 @adamhsparks
I like your solution better.

Send through a PR.
