### 2017-07-16 ToowoombaTrio/bomrang#37 @adamhsparks
Document update_* functions further in vignette
> There's only one "big idea" in the package that I would suggest increasing documentation on. The update_* functions are a cool idea, but I think it would be useful to explain a bit more what's going on with these functions in the vignette. In particular, it would be good if users understand exactly where (in their computer's file system) these updates are being saved. Also, it may be helpful for reproducibility to add some comments on the vignette about how this functionality might result in users with the same version of the package may end up with different results from running the same code (i.e., if one has updated the data at some point and one has not or updated at a different time).

From https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
