### 2017-07-16 ToowoombaTrio/bomrang#38 @adamhsparks
Rename JSONurl_latlon_by_station_name to be shorter
> JSONurl_latlon_by_station_name is a very long variable name

This has been raised by both reviewers.

https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
