### 2017-07-16 ToowoombaTrio/bomrang#39 @adamhsparks
Add minimum version requirements in the DESCRIPTION file
> Some of the package dependencies have gone through a number of big changes since they were first released. It may be worthwhile adding minimum version requirements in the DESCRIPTION file, in case some users haven't updated packages in a while.

https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
