### 2017-07-16 ToowoombaTrio/bomrang#40 @adamhsparks
Refactor code to be more concise
> Some of the code could probably be refactored to be much shorter. One example is the series of checks for whether certain columns are in the data frame within the .get_obs function. Another example is the code to create xmlbulletin based on the state chosen in a couple of the functions, which could probably be shortened a good bit using the case_when function from dplyr. Although, in that second specific case, given that the number of states should stay the same in the future, I guess there probably shouldn't be any maintanance problems resulting from the current approach for that piece of code.

https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
