### 2017-07-16 ToowoombaTrio/bomrang#41 @adamhsparks
Define cook and .get_obs outside functions
I don't think `.get_obs()` is a scoping issue, will have to check, but it could probably be moved to the internal_functions.R file.

> A couple of the functions (cook and .get_obs) are defined within other functions. Is this for some scoping-related reason? Otherwise, it might make the code a bit more readable to define them outside of other functions.

https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
