### 2017-07-16 ToowoombaTrio/bomrang#42 @adamhsparks
Specify conventions for latitude and longitude argument in get_current_weather()
> For the `latlon` argument, it would be helpful to add what conventions the user should use for latitude and longitude (for example, are units decimal degrees? Should southern latitudes be negative, or is that assumed?)

https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
