### 2017-07-16 ToowoombaTrio/bomrang#37 @adamhsparks
Closed with https://github.com/ToowoombaTrio/bomrang/commit/ee5f3e6bc5842cf15e244ca373888cc644b49aa8
