### 2017-07-17 ToowoombaTrio/bomrang#44 @HughParsonage
Wow. This is a really excellent bug report -- I never would have picked it out.
