### 2017-07-17 ToowoombaTrio/bomrang#41 @HughParsonage
I thought it was more readable to put them inside the other functions!
