### 2017-07-17 ToowoombaTrio/bomrang#41 @adamhsparks
I'm opening issues for discussion, we don't have to accept everything proposed during the review if it really isn't necessary or beneficial.

It's worth a look at least for mine, I think, if nothing else to ensure consistency in how the internally used functions are handled.
