### 2017-07-26 #121 @karthik
@adamhsparks Quick update, I'll need a couple of days to read through the comments and look over the changes before providing you with an update. Once I complete that check, I'll check in with Brooke and Mark for their final signoff.

In the meantime you can go ahead and add a ropensci review badge to your README. It will update once this review is complete.
