### 2017-07-27 #121 @adamhsparks
Hi Karthik,
We're still addressing the comments. I'll revert when we're done.

Hope to be done in next two weeks.

--
Adam H Sparks
