### 2017-08-02 ToowoombaTrio/bomrang#39 @adamhsparks
Closed with https://github.com/ToowoombaTrio/bomrang/commit/b33a3087658b27eeadf98ff9845bfc7f3901dc4b
