### 2017-08-07 ToowoombaTrio/bomrang#41 @adamhsparks
I've moved `.get_obs()` outside the main function, `get_ag_bulletin()`, but left it in the same .R file since it's only used in one function and not across the package.

For me, at least, it does help to have it outside the main function I can read that more easily.

I'll close this now and we'll leave the `cook()` function as-is.
