### 2017-08-07 ToowoombaTrio/bomrang#47 @adamhsparks
Use dplyr best practices
> I'm not sure if this has the potential to cause any problems for some users, but it might be worthwhile to try to revise the code to use some of the recommended techniques for standard evaluation when programming with tidyverse packages rather than getting around CRAN notes by defining column names in dataframes as NULL within functions (i.e., all the spots noted now with # CRAN NOTE avoidance). I know that the underscore alternative dplyr functions can do that (mutate_ rather than mutate), but I think they've been deprecated now. I think the latest best practices are outlined here: http://dplyr.tidyverse.org/articles/programming.html

From https://github.com/ropensci/onboarding/issues/121#issuecomment-315419598
