### 2017-08-08 ToowoombaTrio/bomrang#47 @adamhsparks
It took several commits, but this is now done.

The code now uses `rlang::.data`, however, this won't work with _tidyr_ functions yet.
