### 2017-08-08 ToowoombaTrio/bomrang#40 @adamhsparks
Code is refactored for state selection in `get_ag_bulletin()` and `get_precis_forecast()`.

However, since `get_ag_bulletin()` checks for the existence of individual columns, it's not `if else()` statements, but several `if()` statements. I cannot see how this would work in a `dplyr::case_when()` situation.
