### 2017-08-08 ToowoombaTrio/bomrang#38 @adamhsparks
Any thoughts on renaming this? I agree it's long, but it's not something the end user ever interacts with.
