### 2017-08-10 ToowoombaTrio/bomrang#38 @adamhsparks
@HughParsonage, shall I modify the variable name, are you going to or are you opposed to changing it?
