### 2017-08-10 ToowoombaTrio/bomrang#38 @HughParsonage
Hi sorry I'm overseas so can't use GitHub. Whatever you think is best -- no
strong objections here.
On Thu, 10 Aug 2017 at 2:41 pm, Adam H. Sparks <notifications@github.com>
wrote:

> @HughParsonage <https://github.com/hughparsonage>, shall I modify the
> variable name, are you going to or are you opposed to changing it?
>
