### 2017-08-10 ToowoombaTrio/bomrang#38 @adamhsparks
Thanks Hugh. I'll shorten it down.

Safe travels.

--
Adam H Sparks

On 10 August 2017 at 19:01:38, HughParsonage (notifications@github.com)
wrote:

> Hi sorry I'm overseas so can't use GitHub. Whatever you think is best -- no
> strong objections here.
> On Thu, 10 Aug 2017 at 2:41 pm, Adam H. Sparks <notifications@github.com>
> wrote:
>
> > @HughParsonage <https://github.com/hughparsonage>, shall I modify the
> > variable name, are you going to or are you opposed to changing it?
> >
> > —
> > You are receiving this because you were mentioned.
> >
> >
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/ToowoombaTrio/bomrang/issues/38#issuecomment-321465152
> >,
> > or mute the thread
> > <
> https://github.com/notifications/unsubscribe-auth/AHvGDH4C1efZU8oRetXw4ZQs0SxLIhsZks5sWqYUgaJpZM4OZa4X
> >
> > .
> >
>
