### 2017-08-10 ToowoombaTrio/bomrang#38 @adamhsparks
Closed with https://github.com/ToowoombaTrio/bomrang/commit/8bcba0b4c01bcc7410721e285a962f85c87d5200
