### 2017-08-11 ToowoombaTrio/bomrang#21 @adamhsparks
@mpadge, I think we're close to getting all of the point addressed that were raised by you and @geanders during the review process.

Do you want to get started on adding the 9 & 3 bulletins so that we can include this functionality in the JOSS paper submission?
