### 2017-08-11 ToowoombaTrio/bomrang#21 @adamhsparks
We don't have to, I was just offering thinking it would be good to have you on the author list as well.
