### 2017-08-11 ToowoombaTrio/bomrang#21 @mpadge
thanks adam! yeah, that'd be great, but only if 1-2 weeks isn't too long to wait
