### 2017-08-11 ToowoombaTrio/bomrang#21 @adamhsparks
Yeah, no worries, I think it's worth it. :)
