### 2017-08-13 #121 @adamhsparks
@karthik, we're getting there. We've been busy (with this and other things) and we just need to get one more vignette added and then summarise what we've done.

In the meantime, I decided to add some new functionality to the package that the reviewers might want to look at before the final decision. Everything (except tests and documentation, naturally) is in this file for retrieving satellite imagery from BoM - https://github.com/ToowoombaTrio/bomrang/blob/master/R/get_satellite_imagery.R

I need to add a few more tests for this to get our code coverage back up, but I've tested it on my Mac and Linux VMs, so far so good.
