### 2017-08-17 ToowoombaTrio/bomrang#52 @adamhsparks
Add example output to vignette
Vignette lacks examples of output to illustrate what user can expect.

From: https://github.com/ropensci/onboarding/issues/121

> Also, I would recommend evaluating more of the code in the vignette, so that users can see what the output looks like.
