### 2017-08-17 ToowoombaTrio/bomrang#52 @adamhsparks
Fixed with  https://github.com/ToowoombaTrio/bomrang/commit/ee5fd1b83f0f8018566fd0594410e565e83c6e95
