### 2017-08-18 #121 @adamhsparks
Hi all,
Here is a changelog of the updates/changes made to bomrang in response to the reviews.

# Code

## Updates to Code in General

- The package has been linted,
    - line lengths are <80 chars,
    - best practice naming conventions are followed (where possible)

- Onload a message regarding the copyright and data source,
```r
                Data (c) Australian Government Bureau of Meteorology,,
                Creative Commons (CC) Attribution 3.0 licence or,
                Public Access Licence (PAL) as appropriate.,
                See http://www.bom.gov.au/other/copyright.shtml
```
is displayed

- All of the internal files are now loaded from "inst/extdata" and not exposed to the end user and now use
`load(system.file("extdata", "*.rda", package = "bomrang"))` to load the files.

- `agrep` is now used in all functions where the user enters state or Australia values to query BoM data

- best practices for programming with `dplyr 0.7` using `rlang` are now employed, which reduces the need for the `# CRAN NOTE avoidance`

- The DESCRIPTION file now states minimum package versions for packages that are undergoing rapid development

- Code has been refactored to be shorter, _e.g._, `xml_bulletin_url` in
`get_ag_bulletion()`

## Updates to `get_current_weather()`

- "JSONurl_latlon_by_station_name" has been shortened to
"JSONurl_site_list.rda".

- Recommended `station_name` values are separated by spaces in
`get_current_weather()`

- Better consistency in handling names of observation stations for
`get_current_weather()`

- Station names and location names are more consistent in the supplied data and returned data frames.

- Lat/Lon values for `get_current_weather() results are now reported using the values from the internal database, which has a higher degree of accuracy. The json file values are rounded while the values from the stations list has four decimal places

## Updates to `get_ag_bulletin()`

- `.get_obs()` has been moved out of the `.parse_bulletin()` function for easier reading/maintenance

## `get_precis_forecast()`

- fixed a repeat of `return(tidy_df)` in `get_precis_forecast()`

## Additions to code

- Added new, `get_satellite_imagery()` function to download GeoTIFF files from BoM ftp server

# Updates to Documentation

- the _bomrang_ vignette now contains instructions for use along with appendices that document the data fields and units, rather than separate vingettes

- ramifications of updating station lists are now stated clearly in the vignette and help files for applicable functions

- a map of BoM stations is included in an appendix of the _bomrang_ vingette

- Lat/Lon values are specified to be in decimal degrees in `get_current_weather()` help and vignette

- Added new use-case vignette
