### 2017-08-25 #121 @mpadge
@karthik @adamhsparks For the record at this point in the proceedings:

I've added the new functionality that @adamhsparks mentioned above, but it lacks full documentation at the moment. I'm totally happy for the package to be accepted, and suggest that the subsequent JOSS submission could merely be postponed ever so slightly until a more comprehensively documented version can be officially released and doi-ed. An initial CRAN submission with just a lightweight description of the new data should be no problem, and can simply be updated asap afterwards. Sound good?
