### 2017-08-25 #121 @adamhsparks
I'm in no hurry. We can wait until we've got it fully documented to release it on CRAN and submit to JOSS.

It's there on GitHub for anyone that wants to otherwise use it and we might find more bugs in it. I've also a few other issues open that could be nice to address before releasing on CRAN.

Thank you, @mpadge for adding the new functionality. The package is really coming together nicely now.
