### 2017-08-28 #121 @geanders
The package authors did a really excellent job of responding to the comments and suggestions I had. I think the final package is really, really nice. I checked my final approval in the box in my original review.
