### 2017-08-28 #121 @karthik
Hi @adamhsparks,
I'm pleased to accept the `bomrang` package into the ropensci suite. I have also created a team for this package on ropensci's github org. Please accept that team invitation request and proceed with transferring your repo. Once it's transferred, I'll make you full admin and you'll be able to continue updating the package.

**Badges:**

Please add these two badges to your repo:

and the footer:

```
[![](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)
```

Once the transfer is complete you'll need to re-activate the CI services (I will help with this bit).
