### 2017-08-28 ropensci/bomrang#21 @adamhsparks
Closing now as this has been added by @mpadge.

Thank you Mark!
