### 2017-08-28 #121 @karthik
Hi @adamhsparks

Just a last few items to check off!

Since you've asked for a JOSS submission, please generate a new release with the up-to-date Zenodo DOI.

Then please initiate a submission here and suggest me as the editor. I will watch for it and take care of it at the other end.

Before you update the DOI, can you please add a little bit more in terms of research applications for the paper?

I see:

>  The data have applications in applied meteorology and agricultural meteorology and agricultural and environmental modelling.

Also, `satelite` is misspelled in the paper.
