### 2017-08-29 #121 @adamhsparks
Hi @karthik,
It'll be a bit before the JOSS submission as @mpadge is not able to work on this right at the moment and there is still documentation left to add for the `get_weather_bulletin()` functionality before the package is fully ready for submission to CRAN and the paper still needs his ORCID. So it'll be a bit, but I'll add some more detail and correct the spelling in the meantime.

thanks!
