### 2017-09-20 #121 @adamhsparks
Hi @karthik,
I've submitted the latest version, 0.0.7, of _bomrang_ to JOSS this morning and selected you as the editor. @mpadge has finalised the documentation and added a bit more functionality to handle the weather bulletins along with his ORCID and affiliation in the paper.

I've also done spellchecking and other minor fixes here and there to make it nicer.

Thanks for your patience.
