### 2017-10-02 #121 @adamhsparks
I believe so, unless you need anything else from me that I'm unaware of?
