### 2017-10-02 #121 @karthik
Yes this can be closed.
